Massive refactor, converged three columns into one with a slick transformation API.
The goal of this was to improve performance, feature parity, and consistency. There was a rapidly growing disparity in how things like pinning and hiding items was working, especially as each column sprawled out into different trees of code.

I realized that, if each column renders to the same format, there is no point in having different columns. I decoupled the view from the model, and generate the model in a separate piece of code, that creates a shared shape of data from the various BlackBoard models.

This allows me to create complex systems like multiple filters, sorting, and column settings without three sets of basically equal code.